feat(datasource): 升级至 ShardingSphere 5.x 并优化数据源配置
- 集成 ShardingSphere 5.4.1，替换原有 4.x 版本
- 引入 dynamic-datasource 3.6.1 管理多数据源
- 新增百度 uid-generator 专用数据源
- 重构数据源配置类 MyDataSourceConfiguration
- 移除旧版测试数据源相关配置与代码
- 调整数据库表结构，统一使用 bigint 类型存储 ID
- 更新文档说明与架构图，反映最新技术栈
- 通过 AOP 自动切换 uid 数据源，提升代码简洁性

+++ src/main/java/com/lolkt/demo/mapper/OrderMapper.java
@@ -32,7 +32,7 @@ public interface OrderMapper {
     List<OrderPO> selectOrders();
 
     @Insert("insert into t_order (remark,customer_id) values (#{remark},#{customerId})")
-    boolean addOrder(@Param("remark") String remark, @Param("customerId") Integer customerId);
+    boolean addOrder(@Param("remark") String remark, @Param("customerId")  Long customerId);
 
     @Update("update t_order set name =#{name} where id =#{id}")
     void updateOrder(OrderPO Order);
